seemingly working 4.0 grading system awaiting rigorous testing, need to fix floating action button

+++ app/src/main/java/com/engpacalculator/gpcalculator/features/five_grading_system_sgpa_features/presentation/FiveGpaViewModel.kt
@@ -207,7 +207,9 @@ class FiveGpaViewModel @Inject constructor(
                         it.copy(
                             cgpa = String.format(
                                 "%.2f",
-                                _fiveCgpaUiState.value.cgpaList.sum() / _fiveCgpaUiState.value.cgpaList.size
+                                (_fiveCgpaUiState.value.cgpaList.sum()
+                                    .toDouble() / _fiveCgpaUiState.value.cgpaList.size.toDouble())
+
                             )
                         )
                     }
